requirements.txt
django==4.2.13
djangorestframework==3.15.1
django-cors-headers==4.3.1
dj-database-url==2.1.0
whitenoise==6.6.0
cryptography==41.0.7
reportlab==4.1.0
pandas==2.2.2
sslyze==5.2.0
dnspython==2.6.1
requests==2.32.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
networkx==3.3
apscheduler==3.10.4
python-dateutil==2.9.0
pytest==8.2.0
pytest-django==4.8.0
gunicorn==22.0.0
